Add some checks for mesh, improve edge vertex search
I was confused for a moment on the "No edge vertex found" warning
suddenly displayed by Strata. This turned out to be due to Stitches
being made but not initialized, and then subjected to
fixSearchParameters().

I decided that some more checks on meshes being 'complete' were in order
(the incomplete stitches had 1 vertex each). Also, while looking around,
I noticed that TopologicalMesh::findRandomEdgeVertex() contained a bug:
it looks for vertices in the index array 've', but those are not
guaranteed to exist. Although invalid (non-vertex) entries of 've'
should always have the error value of 0, and nothing should break, it is
still bad style to traverse 've' and thus I adjusted the function to use
'vertices' instead.

This commit comes after a bit of hiatus that I blame on Blade and Soul
coming out. "Fortunately", however, I injured my finger so no
fast-paced gaming for me for a while. Strata, here I come...