Add functions for determining postioning near a mesh
I am not done yet, but nonetheless I am a few steps closer to 'finding'
the underlying vertex of a point. As I continue implementing the
required algorithms, it occurs that some more subtle treatment is
required, and hence the complexity of the functions involved grows.

I realised that for a good determination of the correct underlying
vertex one must be able to properly compare several candidates. This
requires a topological function for determining whether one candidate is
from a layer that overlies the other candidate.

This function now appears to be the only thing missing. I have proceeded
to the point where everything else is at least more or less in the
required form, even if it is untested and possibly incomplete.

Therefore, to be done is writing the function Bundle::isNearMeshAtVertex
and once that works, checking that the correct vertices are found
indeed. Then, the actual stitching of the layer onto the underlying
terrain can be performed. That stitching should merely consist of adding
some functions for traversing the underlying vertices, and thus we
should be able to make use of the functions being added now for finding
the initial underlying vertex.