Isolate Bundle::findRemoteNeighborVertex()
This commit separates the function for finding a neighbor vertex from
among all the neighboring Strip objects.

The reason is that this functionality is far too useful, and I need it
for isAbove/BelowMesh functionality as well. I want to avoid too much
duplication, therefore I took the core of the search algorithm and made
it into a standalone function that can be used for both purposes.

The only change is that one can now choose whether to look clockwise or
counterclockwise. This is required to find all neighbors for vertices
that are at the Layer's edge (which, by their very nature, do not allow
for a circumnavigation along all neighboring vertices).

I am not yet done with the above/below functions so that part remains
commented out for now.