Calculate mesh normals regardless of mesh divisions
There was already a method for calculating the normal as an average of
all the normals of the polygons that a Vertex belonged to, but that
calculation did not incorporate polygons outside of the Mesh for which
the calculation was done. The functions added here
(Strip::computeSumOfPolygonNormals and Bundle::CalculateVertexNormal)
perform a more general version of this. I really wanted to have this
general version, because it is independent of the way in which the mesh
is cut into Bundles and Strips. Such independence seems important for
getting accurate results, and that is what I need to ensure that finding
the underlying vertex always results in a vertex that really lies on the
layer directly underneath.

I also worked a bit more on Bundle::isNearMeshAtIndex() but I am not
done. I need to think about a safe way for determining a 'cone' from a
set of normals, in such a way that points in space are guaranteed to fall
inside at least one cone for a mesh, no matter where the points are.
Given that the mesh can have all kinds of shapes and that the 'cones'
are really defined by squares, not triangles, that seems to be
guaranteed intuitively but I am not entirely convinced - yet. And the
'inside' also needs a careful definition in terms of vector algebra.