Finally found the underlying vertex
It has been quite a long way, but... it appears that I have finally
found the way to determining the correct underlying vertex.

The main issues were considerable subtleties in defining what it means
to be 'near' a mesh, as well as a previously unnoticed couple of
weaknesses in findIntersection(), the function that does line-plane
intersection. And then there were the usual silly mistakes with writing
the wrong variable and taking pretty long to find out why the results
were so strange.

With the underlying vertex found, next up is connecting the underlying
vertices and forming a Stitch during the process.