De-const findRemoteNeighborVertex c.s.
Today I probably should've gotten farther, but at least some issue got
the little bit of consideration it deserved.

That issue was the current way in which the Bundle and Strip are finding
neighbor vertices. As it is a search function, it was so far const-ed -
as it should not really change the class, but merely look around.

That changes now, however. I need to stitch. I do not only need to look
for the neighbors of vertices, I need to actually build them into
StripVertex objects and form Stitches from them. It isn't possible to
make StripVertex objects from const references to vertices - the whole
construction of meshes relies on strips and bundles being able to remove
their respective remote references from each other.

I therefore thought for a while, discarded the idea of mapping const to
non-const (since that would be even uglier), and accepted the
inevitable. Now, all kind of search functions are non-const, since some
parts of the functionality need to return non-const references to Bundle
objects, and I prefer non-constness to widespread code duplication.

As far as the actual stitching is concerned, I barely made any progress
but at least the next time this issue won't be there. The de-consting
allows adding the only actual code line added - namely, finding a
neighbor along the layer's edge. This was done by basically stealing the
code from Bundle::isAtLayerEdge() and make the latter function a special
case of the more general who's-my-neighbor function,
findAlongLayerEdge().

That finds the next stitch vertex for the upper layer. As for the
underlying terrain, a function must be added to select the most suitable
neighbor of the lower layer's leading vertex, using its proximity to the
upper layer's leading and next vertex. That will again need a bit of
geometrical trickery.

Finally, a bit of cleaning up of Bundle::findRemoteNeighborVertex() is
warranted. It should probably use StitchVertices - errr, StripVertices
(this spelling confusion keeps plagueing me), and then only one (the
current situation with a neighborBundle and nextBundle both being passed
as arguments is awkward).