Restructured Bundle::findRemoteNeighborVertex() et al
As promised a long time ago, this commit cleans up the rather clumsy
structure of the across-the-Layer topological funcionality in the Bundle
and the Strip. The way things became like that is basically that a
(working) approach was split up several times, with eventually too much
complexity surviving. In addition, my initial approach was created
before I started using the StripVertex as the natural pair object to
contain (Bundle*, xVert) pairs.

With the latter approach increasing in usage, mostly thanks to the nasty
confusions it avoids, it was time to change things here as well.
However, adjusting the code took a couple of hours - far longer than
reasonable - mostly because the code was so subtle and so impractical.

Now, however, StripVertices are used for all Bundle-index pairs, and lo
and behold: the code is much shorter, and also much more transparent.
The true beauty is underscored by statements such as

 while(neighbor != endVertex)

which is replacing the far less transparent condition

 while(neighborIndex != endIndex || neighborBundle != this)

in Bundle::findAlongLayerEdge(), or equivalently

 if(vertices[i] == v)

replacing

 if(vertices[i].getRemoteIndex() == v && vertices[i].getOwningBundle()
== vBundle)

in Strip::findRemoteVertexPolyNeighbor(). Such statements employ the
StripVertex's equality and inequality operators, encapsulating the
logic of whether two Bundle-index objects are equivalent (within the
Layer's scope) or not.

In fact, it's a pity Strata isn't actually used by anyone yet, since it
would be awesome to share the sheer beauty of this implementation with
others. So for now I will merely be very content with my improvement.

On a general note, the time passing since the last commit is manifold: I
have played a bit more of both BnS and Asta since my finger got better,
I have a lot of people in my house lately, and early on my finger didn't
exactly approve of extended typing (but fortunately that's a lot better
now). I do intend to continue, though the pace may continue to be
disappointing. The idea is to get to a working Strata as soon as
reasonably possible (with 'working' defined as being able to safely
manipulate gapless terrains), and then focus a bit more on adding
actual functionality. I aim to do folding and subduction first, since it
seems easy, followed by erosion and finally sedimentation.