Clean up obsolete comments from Bundle and Strip
After the recent work on mostly Bundle (but also Strip) there was a
considerable amount of code that I wanted to comment out first (in
earlier commits) to make it easier to compare the changes made to the
code. However, the original code will never be used anymore, and a lot
of comments relating to it is no longer applicable.

Therefore, this commit only cleans up unused code, and it does not alter
functionality.