Track Layer's edge across underlying Layers
This commit brings the next stage of the Stitching functionality:
finding vertices on the underlying terrain (of a Layer to be stitched)
that are adjacent to an already existing Vertex (of the stitch), and in
the general direction of the upper Layer's vertex.

The code is not yet perfect - I still have to think of fixes for harder
cases, such as more complicated geometries (where the underlying terrain
is strongly curved or contains layer edges) - but since it works for the
base case, it is a good point to commit the current situation.

Next to be done is ensure that the selected neighbor vertex is never
underlying to another neighbor vertex. If that is the case, there must
be a Layer edge and the Stitch being created is at risk of following a
Layer that is not on top of the terrain. This situation does not occur
yet, but in the future it will almost certainly become an issue if it is
not anticipated now.

After correcting for the above, the focus should then shift towards
forming Strip polygons for forming the Stitch that connects the Layer to
the underlying terrain.