Check above-mesh-ness for lower Stitch vertices
In this not-so-micro-commit I add a preliminary fix for the mentioned
issue that I anticipated: namely, that it is hard to ensure that
stitching is always done to the uppermost Terrain layer, since the
nearest Vertex to a given point is not necessarily a vertex at the
surface.

The solution is then to check whether or not the candidate vertex is
'above' the other candidate (or not). If one of the vertices is above
the other, then clearly the above one should be used regardless of who
is closer.

On the other hand, if there is no above-and-below ordering, then the
distance should determine the location. This situation leads to a
three-way check for determining whether or not to replace the existing
candidate with a newly found one:
- if the new candidate is above the old, then use it;
- if the new candidate is a neighbor of the old and closer, then use it;
- and if the new candidate is not a neighbor, and not under, and closer,
  then use it.

The check is admittedly complicated, but given the disastrous results of
not properly following the surface (by all means one might never manage
to even return to the initial point!) it is quite necessary to develop a
method that is failsafe enough to work in all but the most horrible
scenarios.

Should the current approach still not work (e.g. due to the initial
underlying vertex not being on the top layer after all) then there
should be some effort to define being-on-top more clearly, e.g. by
calculating the normals of all polygons adjacent to a vertex whenever
the vertex is related to a Stitch (signalling that one layer is covering
another at that place). With such a check one could establish beyond
doubt which neighboring vertices are at the surface.

Such logic is however not available at this time and it is not possible
to use this approach to locally determine the top layer (since one can
only categorize neighbors, and not compare location relative to vertices
on other layers). Therefore the current approach is retained for now.