Fix issues in TopologicalMesh::findNearestNeighbor()
While the previous commit did compile properly, it did not always find
the correct vertex as the closest underlying vertex to the leading upper
vertex. It turns out that was due to sloppy checks in
findNearestNeighbor(), which only compared the current nearest candidate
to the two neighbors of each polygon, without comparing these neighbors
to each other. In situations where the neighbor existed in only one
adjacent polygon (quite common for at-Layer's-edge polygons), the
closest neighbor was therefore never considered as a candidate.

This commit thus improves the candidate selection logic for intra-mesh
nearest neighbor searches. For the current parameters, the correctness
of nearest-neighbor searching is confirmed by the distance between the
upper and lower leading vertices being 2 (i.e. the Layer's thickness)
everywhere.