Merge RemoteVertex into StripVertex
In this small commit I transfer the existing usage of the RemoteVertex
to the StripVertex (through the use of a temporary typedef).

Functionally there is a subtle difference: the RemoteVertex could not
access the remote vertex's position and therefore had to be constructed
with the position as one of its arguments (even though it would
eventually be used to create a StripVertex). The position was important,
though, as it was used to check whether a vertex was already added to
the Mesh at that position.

With the addition of StripVertex::resetPosition() that complexity is no
longer explicitly required, and thus the code has become somewhat
simpler as well.