Rename StripVertex -> RemoteVertex
In this commit I finalize the adjustment to merge the StripVertex and
the RemoteVertex into a class with the functionality of the StripVertex
and, well, the name of the RemoteVertex.

With this, the road should be a little more clear to finally implement
Terrain::stitchLayerTransverse(), which already saw some expansion two
commits ago and which should be the primary target for the upcoming
commits.