Remove unused StripPolygon class
The StripPolygon was a class once devised to represent polygons
consisting of StripVertex objects (recently renamed to RemoteVertex),
but there is no reason to believe there will ever be a need for these
special types of Polygon objects.

In the earliest stages, the polygon type of the mesh used to be a
template paramater PolygonType or so, similar to the VertexType being a
Mesh template parameter.

However, there turned out to be little reason for the distinction, and
thus Strip objects use normal Polygons. Given Strata's gradually
advancing degree of complexity, it seems reasonably certain that this
will remain as it is now, and there is certainly little reason to hold
on to an unused class for much longer.