Remove obsolete RemoteVertex constructor
In this commit I removed the RemoteVertex constructor where the position
is specified.

This constructor was initially the default one, as positions could not
be passed, but since resetPosition() has been added this is no longer
necessary. Since it is, in any case, quite inconvenient to have to
specify the position all the time (information which is already uniquely
determined by the Bundle*-xVert pair being passed), as well as dangerous
(considering the kind of issues that lead to resetPosition being added),
I thought the time was right to get rid of it.

Fortunately, it turns out nearly no one was actually using this
constructor directly (with mostly everyone already using the
position-less constructor), and so this commit is pretty insignificant
in terms of lines changed.