Improve stitch step condition
This small commit improves the (yet unused) logic for determining how
the stitching of a new layer onto the underlying terrain should
progress.

As stitching is defined through progressively finding new vertices for
both the upper and the lower layer, there must be a loop through which
polygons are added iteratively, and both the upper and lower layer
vertices are iterated over.

The current plan is to prefer adding the polygon that is most far-behind
in the iterative procedure, with far-behind being defined according to
both the upper and lower iterators (where each iterator consists of two
vertices, one leading and one trailing). More precisely, the upper
iterator will move (and the lower iterator remain) if any of the
following is true:
1. the lower trailing vertex is far ahead of the upper trailing vertex,
   such that it is closer to the upper leading vertex than to the
   trailing one;
2. the lower leading vertex is far from the upper leading vertex,
   farther than the lower trailing vertex is, such that we cannot find a
   good new lower leading vertex;
3. the leading vertices are both well ahead of the trailing vertices,
   but the vertex formed from 2 upper vertices has shorter, more
   proportional edges.

In the remaining cases, the lower iterator is moved. This procedure is
largely symmetric, but the main difference is that after the upper
iterator moves, we need to re-check if the lower iterator is still
properly positioned. In particular in case (2), but also in the other
cases, it is quite likely that a new upper leading vertex (which is the
target to be followed for the lower iterator) has a different neighbor
of the lower trailing vertex than the present lower leading vertex as
its preferred lower leading vertex.

This commit does not yet perform the re-evaluation of the ideal lower
leading vertex, but it is intended for the subsequent commit.

I also do not have anything in place for finishing the loop yet, so I
anticipate that there will be a case (4) where the upper iterator
increments if the lower iterator can increment no more because the
circle is complete.