Create Stitch meshes to connect Layers
After a considerable delay, here I finally commit the one result that
the last thirty or so of commits (for the past half a year) were after:
namely, to stitch layers onto the underlying terrain.

With the previous commit being so close, the finishing touch was more
demanding than I had hoped for. This is due to several emerging issues:
- Walking along the Layer edge and finding remote Neighbors did not
  correctly incorporate (non-)clockwiseness;
- The requirement for the third condition of the is-a-nearby-vertex
  check also incorporated that the new candidate is a neighbor of the
  old one, which is too strong;
- The stitching looked for leading upper vertices in the
  counterclockwise direction while stitching was intended to proceed
  clockwise;
- The stitching always tried to increment the upper vertex while the
  lower vertex was at its starting point.

Issue 1 was fixed by adjusting Bundle::findRemoteNeighborVertex() and
Bundle::findAlongLayerEdge(). Issue 2 was fixed by commenting away the
neighbor requirement in the third condition seen in the large
if-statement of Bundle::findNearestNeighborInBundle(). Issue 3 was fixed
by changing 'true' to 'false' as the last argument to both calls to
Bundle::findAlongLayerEdge() in Terrain::stitchLayerTransverse(). Issue
4 was fixed by requiring that both upper and lower were at their
starting positions (rather than just the lower one).

The actual order in which the issues were discovered is actually
4-3-1-2, and in fact there was already some 'functional' stitching
pretty quickly (i.e. after fixing issue 4), but that stitching would
follow Bundle edges rather than Layer ones because issue 3 combined with
issue 1 somehow resulted in that behavior. The fact that it almost
worked somehow made it a lot harder to find out that the infinite loop
resulting from issue 2 was actually there. Basically by fixing 1 and 3 I
broke it all and it took me a while to realize that there was something
else wrong that somehow wasn't apparent initially by pure chance.

But of course that's just a programming lesson for me, and hardly
anything of true relevance for the commit. I am glad that we can finally
leave the stitching behind and continue to implement other fun elements
of Strata.