Add proper adjacency tracking to Stitches
Now that the Stitches work, it was about time to ensure they properly
tracked all their adjacent Bundles.

It turns out this was pretty subtle. Firstly, when I added another Layer
on top (such that I need to Stitch onto a Layer that already has
Stitches, and such that I tested duplicating a Layer with Stitches) I
noticed that Stitches were not excluded from duplication everywhere. I
therefore adjusted Bundle::duplicateBundle() and
Bundle::duplicateAdjustAdjacentStrips().

Elsewhere, for proper Layer edge tracking, I had to distinguish between
whether or not to include Stitches when finding (possibly remote)
neighbors. I therefore added a 'skipStitches' flag to both
findNearestNeighborInBundle() and findRemoteNeighborVertex().

In the Strip, I disabled the check for non-Stitch vertices in Stitch
meshes because the criterium it used (that Stitches consist of polygons
where at least one RemoteVertex has a secondaryOwner, i.e. it is a
stitch vertex) is too strict. It is quite permissible for Stitches to
contain Polygons that each depend on only a single remote vertex.

In the TopologicalMesh I made some topological checks significantly
slower (but safer) after I ran into infinite loops. Apparently the usual
disk check (where we loop through all of a Vertex's neighbors) can go
horribly wrong when the topology around a vertex is messed up. I
therefore added two tests: whether a neighbor is seen more than once,
and whether the total number of neighbors exceeds
STRATA_VERTEX_MAX_LINKS (i.e. the hard-coded neighbor count limit). This
also determines the nature of the issue, but it has the advantage that
it manages to avoid getting stuck for eternity.

That being said, there is apparently something wrong in the Strip that
connects the top layer to the middle layer, since it generates errors
for two of its vertices, while it definitely ought not to do that.
However, Strata does run kind of fine and the stitches don't look very
bad, so for now the adjacency is committed.

The adjacency itself is done through adding the Bundle as an adjacency
for every vertex that is added to the Stitch. This may seem a bit
excessive but I do not see any obvious way of being more efficient. And
it shouldn't be a big performance hit, anyways.