Fix issue with selecting poor underlying vertices
The issue causing all the misery yesterday, with the topology checks
failing on the Stitch of the second layer, was indeed a sign something
was deeply wrong.

In two instances, it turns out that Stitch polygons were still in the
equation while they should not be. Firstly, they were involved in
calculating the normals of vertices. This is unnecessary, since using
non-stitch polygons only already gives a normal perpendicular to the
layer's surface, but for transverse stitches (such as the ones we use
now) it is also wrong: at the edge of the Layer it makes the normals
point tilted outwards, even though in principle the layer's surface
normal is still vertical. This issue was addressed by skipping adjacent
strips in Bundle::calculateVertexNormal() when they are stitch Strips.

Secondly, I erroneously passed 'false' to skipStitches in the call to
findRemoteNeighborVertex() made from isNearMeshAtIndex(). That was a
 mistake: when adding skipStitches I put 'false' in two places in
bundle.cpp that made calls to findRemoteNeighborVertex(), and 'true'
everywhere throughout terrain.cpp where I needed
findNearestNeighborInBundle(). In hindsight, these should all have been
'false', and indeed typically one should pretend that Stitches don't
exist at all - they merely serve to stitch stuff together, but they
aren't really part of the Layer in the way Bundles and non-stitch Strips
are.