Clean up RenderManager; split screen/worldRenderable
In this commit I split the renderable counter that assigns unique keys
to Renderable's into two counters: one for the worldRenderables and one
for the screenRenderables. We will also use only one key map for the
Renderables.

The general reason for doing things about the RenderManager is that I
want to have console-like functionality in Strata and for that I want to
have a better functioning window system. Technically this
adjusting-the-counter wouldn't be necessary yet, but it would have to be
done at some point anyways, and I do need to make the RenderManager more
complex. So I decided to start by improving the things that are already
there, before adding new stuff.

A long time ago I forked the tiny-game-engine and restructured its
method of rendering Renderables, which in principle made it possible to
render things in a different order than they had been added. Now seemed
like a good moment to actually start implementing functionality that
allow windows to be in front of each other, without having to completely
delete and re-add all objects that need to be in front of something
else.

Consider the following: there is an active window in front of the world,
and at some moment something happens that requires a new screen element
to be added that has an effect on something far away (e.g. a new light
source). Then we would have to first remove all the windows, add the
light source, and then re-add all the windows. That would be fairly
ridiculous.

However, now we could in principle allocate 'blocks' of keys to certain
layers of the UI. We could - for instance - reserve the first 1000 or so
keys for all the most-faraway things, then the next 1000 keys for
non-active windows, then 100 keys for the top window and its attributes,
and then another 1000 keys for drawing the primary UI elements on top of
everything else (e.g. selected parts of the HUD such as the primary menu
toolbar).

The actual block reservation system has yet to be developed and I'm not
even sure if that is what I will do next. After all I don't think it's
necessary for the foreseeable future - but one day it will be.