Use square texture as background; start interpreting input
This commit finishes the partial-square-texture that establishes a
method to draw shaded backgrounds for UI windows. It requires an
adjustment to the tiny-game-engine, currently a pending pull request
named partial-filling-screensquare, as the original tiny-game-engine
only permitted screensquares that covered the whole screen. Meanwhile it
continues to require the pending pull request
typecluster-const-iterators. It is a bit of a hassle to get all the
required functionality, but on the other hand as long as it's only me
who needs to compile Strata it's something that I can live with.

After the screen squares, I made a bit of a start with the
InputInterpreter, a class in strata::ui that I intend to build in order
to regulate user input.

The idea I am having is still vague, but goes as follows. User input
must be dealt with in contextual ways: what certain input does depends
on the current state of Strata. For instance, during console mode the
typical rotations should be disabled. Additionally, it should be
possible to redefine key bindings, so the InputInterpreter should
(eventually) be able to support a key remapping that maps standard keys
onto customizable keys (in fact, it's more the other way around, but
anyways).

The key bindings are being left untouched for now, while I focus on the
contextual input binding. I made a UIListener in the UI interface that
allows UI elements to subscribe to certain input, and to receive
subscribed inputs. In practice, I suppose that subscription will be
strongly dynamic, e.g. the console can be subscribed to 'Enter' only by
default, but once 'Enter' is pressed it will also subscribe to all
alphanumerical characters so that one can type normal sentences.

I will probably have to define sets of characters for ease of mass
subscription. Additionally, I will have to think about how to deal with
the inevitable spaghetti that results in a large program where many
elements will want to subscribe to input depending on a large number of
factors. Since I cannot presently see exactly what that will look like,
I will aim to be as careful as possible in allowing contextual
subscriptions, probably restricting them to be input-based only, and
perhaps universally binding 'Esc' as the input that kills existing
subscriptions.