Create InputSet and start defining key sets
For input classes to be conveniently subscribed by UIListeners, in this
commit I add the InputSet (as a set of characters to be subscribed under
a certain set of rules) and a couple of key sets (in
interface/uikeys.hpp) that can be used as standard collections of
SDLKey's.

I also noticed that in SDL 2 the SDLKey works pretty much different, but
after some consideration I decided to carry on as planned anyways. The
tiny-game-engine uses SDL 1.2, so for Strata's current purpose there
isn't really another way, and in any case I only use the enum to define
the input - if there is another way of receiving input in the faraway
future, it should not be a big deal to convert it into the SDLKey enum.
If you wish, it should be sufficient to think of the SDLKey as merely a
large set of possible triggers that Strata can receive - how exactly
they are received (through keyboard presses reported by SDL through
SDLKeys or otherwise) is not really critical.