Move interface/uikeys.hpp -> interface/keys.hpp
Since I am a lazy... well, person, I am moving the 'uikeys.hpp' in
/interface/ to 'keys.hpp'. The main reasoning why the move should not be
done is that it doesn't really matter (the content makes the function
clear and in any case the filename seems sufficiently descriptive either
way), but besides the lack of reasons not to do it, the reason to do it
is obviously that I keep getting annoyed with the tab-autocompletion
that wants to know whether I want to go to 'ui.hpp'/'ui.cpp' or
'uikeys.hpp'. Well, that little problem should be fixed now... in what
is probably the least useful commit to Strata in a pretty long time.