Implement InputSet and InputInterpreter
This commit concerns the core structure of input handling by the
InputInterpreter. The InputInterpreter now keeps a list of InputSet
objects that define subscriptions to input by various parts of Strata.

Currently no classes are UIListeners, and the added code is not yet
used. Examples of future UIListeners include a to-be-defined main menu,
a console allowing Lua commands, and a rewritten exploration mode
(currently moving around is not exactly elegantly implemented, as it
still uses the tiny-game-engine's updateSimpleCamera).

I spent considerable time changing my mind on the data structure of the
InputInterpreter's list. I started with a set in order to keep
uniqueness easy, but then I couldn't make it an ordered list. I moved to
a deque to keep the ordering clear, but realised while writing that I
rarely ever needed random access and that reordering the list (as well
as adding/deleting) was messy, and that basically there was no good
reason not to use std::list (other than that iteration becomes slightly
less elegant as it uses iterators rather than integers).

I made the three functions that one needs for a functional input
mechanism: subscribe(), unsubscribe(), and bump() where the last one is
the only option that allows one to acquire priority on given input
without having to adjust someone else's InputSet.

If that looks strange, imagine the working of a console. When someone
brings up the console, input should no longer affect the rest of the
game, but when the console goes it should again affect everything. Only
the console itself knows whether it's up or not. It would be awkward to
start messing with everyone else's input registrations simply because
the console is (de)activated. Hence the addition of the bump() option.

The obvious consequence is that every listener that bumps itself must
ensure that there is always a clear way to un-register most of its
defined input. In fact, it is best to unregister *all* input for such a
listener, and only let the listener subscribe to input (and be bumped)
by another listener (that is not bumped). In other situations, the input
registered for a bump-able listener should be globally unique (i.e. not
defined by any other listener except for limited duration by another
bumpable listener).

In the future it may become convenient to allow 'relative' bumps where
listeners are moved relative to other listeners. I do not know at this
time whether that will ever be used, so I have not implemented it.

Small changes:
I adjusted the keyDown/keyUp functions to a generic keyInput function,
since that saves one pure virtual function to be implemented by every
UIListener. Unfortunately the tiny-game-engine does not do this so there
are still two functions for this in the ApplManager.

Finally I realised that there is a fundamental shortcoming in the
tiny-game-engine, as it does not support accessing the SDLMod part of
the key event, such that one cannot determine what modifiers were active
at the time of the key press. For now, it doesn't really matter, but in
the future (e.g. if Caps Lock is going to be supported) it could become
a problem. I will try to adjust for this shortcoming (at least
partially) in the next commit. I also created Issue #16 of the
tiny-game-engine to start a discussion on how to fix this.