Include SDLMod in input interpretation
Due to the tiny-game-engine's inability to transmit SDLMod information,
initially I only worked with SDLKey objects. However, I realised it is
not all that hard to include the modifiers anyways, since
SDL_GetModState() allows getting the current modifiers (which, if done
after keyDownCallback() is called, supposedly gives the modifiers that
currently apply). I am not certain whether it is universally true that
the modifiers are the same as those normally contained in the 'mod'
field of the SDL_keysym object, but I hope it will be close.