Allow hiding/unhiding the Monitor
In a first step towards interpreting input, I implemented a proof of
concept that allows one to hide and show the Monitor window.
- Hide: Ctrl-q
- Show: m

The main question here was whether closing a window should cause the
window class to be destroyed, or simply to make it invisible. I decided
in favor of the latter, since destroying and re-creating is not quite
convenient and since rendering windows is in principle not that much
effort. Anyways, the window *is* blanked (i.e. no text is rendered, not
even invisible text) so the performance cost remains minimal.

More broadly speaking, in all likelihood it will eventually be necessary
to destroy entire UIListener objects. The current state of Strata is not
sufficiently advanced for that, since it entails the ability to create
certain Lua-defined objects dynamically (not only on startup) - after
all, the properties are eventually supposed to be Lua-determined.

That being said, it was never the idea that all Lua information was
fully accessible at runtime, and it is quite acceptable to have some
elements be loaded through Lua definitions on startup only.