Add Main Menu; restructure Window
In this commit I added a functional Main Menu which replaces the
original method of exiting Strata (namely by quitting on the UIManager
receiving a SDLK_ESCAPE key input). This ensures that key input is
properly processed and prioritized.

The Window class now does a far larger share of the work, with some
functionality moved into there from the Monitor/MainMenu classes. The
Window handles itself whether it is visible or not, it gets a first look
at the key input (which is forwarded to the derived class if the Window
doesn't use it for opening or closing), and it now contains a title, a
secondary colour, and a drawTitle() function.

The main menu is now also implemented in Lua. I added a second colour
scheme to the Window object in Lua, which can be used for highlighting
to emphasize parts of the text. Further colours should not be part of
the base window class, but be specified via a to-be-added text encoding
(e.g. using HTML-like tags). This is primarily because one cannot really
implement every possible colour variation; it must be Lua-specified (but
also without needing a separate Lua variable for each piece of coloured
text).

The larger Window class also made it practical to adjust the Lua
function signature for window base info (also the UIWindowBase class is
as much bigger). That said, the only thing else that I currently want to
specify through the Lua function is the box frame texture (i.e.
background and borders), so it shouldn't get too much longer.

With these matters addressed, there is still room in this area for at
least two more major additions. Firstly, I want to add the console that
was the main reason for starting work on the UI. Secondly, I want to
specify textures through Lua, eventually by referring to texture
locations but for now by making use of the default texture creation
(where Lua could still choose the hue).