Load backgrounds from Lua; generalize attributes
This commit brings a generalization that makes me happy: namely; to pass
attribute-value pairs to windows that are unpacked by the final window.
That was necessary because I started to dislike the abundant growth of
(1) windows and (2) arguments to Window creation functions in the
UIManager.

When we try to parametrize windows in Lua, one can of course pass the
same kind of parameter to every single window, even if every window has
the parameter, but that defeats the advantages of the class hierarchy
that is the Window. Instead, every Window property should be sent from
Lua to the target Window without the UIManager needing to be explicitly
involved in the transmission of the parameter.

Said otherwise, the UIManager should simply transmit the message, and
not need to know the contents. Moreover, the UIManager should in fact
not be concerned about specific Window types at all. It should simply
create them - on Lua's request - and then treat all Windows equal. If
Windows are truly encapsulated, then that means they should be perfectly
able to take care of themselves.

Therefore, the initial situation where the UIManager has a pointer to
the Monitor and another one to the MainMenu, and needs to update them
explicitly one by one, is not necessary at all. Instead, the UIManager
can simply list Window's and update them all in the same way, and the
Window objects can use polymorphism and proper InputSet's in order to
communicate with the rest of the world without the UIManager being
involved in it all.

I am convinced that this is the way to go, so I moved the fps function
to the ApplManager, and I ensured that no Window is updated with 'dt'
being passed as an argument. This seems reasonable anyways, since a
generic Window doesn't care about what time it is.

Besides the generalization of Windows in the UIManager class, I also
added backgrounds. The above improvements basically came into being as I
got distracted while trying to load background colors from Lua, and I
realised that the current direction felt wrong.

Finally, this is not the end of this chapter: most of the arguments
still being passed to create the UIWindowBase can perfectly be passed as
attributes, too. This will hopefully be addressed in a future commit.